add config: setting interface is optional. RIPE probe disable it
Disable interface name feature, if you don't need it.
interface name only works with patched libevent.
The libevent patch is not in the upstream.
And it is less likely to be upstreamed because it uses a Linux specific code

RIPE atlas WiFi probe need this to disable with libevent patch.